requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
# bzip2=1.0.8=h99b78c6_7 # Conda specific
# ca-certificates==2025.4.26 # System/Conda managed
certifi==2025.4.26
charset-normalizer==3.4.2
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
interchange==2021.0.4
javalang==0.13.0
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langchain-upstage==0.6.0
langsmith==0.3.42
# libexpat=2.7.0=h286801f_0 # Conda specific
# libffi=3.4.6=h1da3d7d_1 # Conda specific
# liblzma=5.8.1=h39f12f2_1 # Conda specific
# liblzma-devel=5.8.1=h39f12f2_1 # Conda specific
# libsqlite=3.49.2=h3f77e49_0 # Conda specific
# libzlib=1.3.1=h8359307_2 # Conda specific
marshmallow==3.26.1
monotonic==1.6
multidict==6.4.4
mypy-extensions==1.1.0
# ncurses=6.5=h5e97a16_3 # Conda specific
neo4j==5.28.1
numpy==2.2.6
openai==1.78.1
# openssl==3.5.0 # System/Conda managed
orjson==3.10.18
packaging==24.2
pansi==2024.11.0
pgvector==0.2.5
pillow==11.2.1
pip==25.1.1
propcache==0.3.1
psycopg2==2.9.9
py2neo==2021.2.4
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
# python=3.12.0=h47c9636_0_cpython # Conda specific
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
# readline=8.2=h1d1bf99_2 # Conda specific
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
# tk=8.6.13=h5083fa2_1 # Conda specific
tqdm==4.67.1
tree-sitter==0.24.0
tree-sitter-java==0.23.5
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
# tzdata=2025b=h78e105d_0 # Conda specific
urllib3==2.4.0
# wheel=0.45.1=pyhd8ed1ab_1 # Conda specific
# xz=5.8.1=h9a6d368_1 # Conda specific
# xz-gpl-tools=5.8.1=h9a6d368_1 # Conda specific
# xz-tools=5.8.1=h39f12f2_1 # Conda specific
yarl==1.20.0
zstandard==0.23.0
